refactor(tests, scripts, credentials): update PocketbaseAuth imports and cleanup
- Updated `PocketbaseAuth` imports to use the Common directory for consistency.
- Removed redundant log verification steps in `full-test.sh` script.
- Improved script messaging for test execution clarity.
